Refactor And ServiceEmprunt Working with Synchronised

+++ src/BD/ConnectionBD.java
@@ -2,14 +2,15 @@ package BD;
 
 import model.Abonne;
 import model.DVD;
-import model.Document;
+import model.Etat;
+import model.IDocument;
 
-import javax.swing.plaf.nimbus.State;
 import java.sql.*;
-import java.time.LocalDate;
-import java.time.format.DateTimeFormatter;
+import java.text.DateFormat;
+import java.text.ParseException;
+import java.text.SimpleDateFormat;
 import java.util.ArrayList;
-import java.util.Locale;
+import java.util.Date;
 
 public class ConnectionBD {
     private Connection conn;
@@ -29,39 +30,63 @@ public class ConnectionBD {
         }
     }
 
-    public ArrayList<Abonne> getAbonnesFromDB() throws SQLException {
+    public ArrayList<Abonne> getAbonnesFromDB() throws SQLException, ParseException {
         ArrayList<Abonne> abonnes = new ArrayList<>();
         String requete = "SELECT * from ABONNE";
         Statement req1 = this.conn.createStatement();
         ResultSet rs = req1.executeQuery(requete);
+        //convertir la date de naissance au format  YYYY-MM-DD en Date
+        DateFormat dateFormat = new SimpleDateFormat("yyyy-MM-dd");
+
 
         while(rs.next()) {
-            abonnes.add(new Abonne(rs.getInt(1), rs.getString(2), rs.getString(3)));
+            String dateNaissanceStr = rs.getString(3);
+            Date dateNaissance = dateFormat.parse(dateNaissanceStr);
+            abonnes.add(new Abonne(rs.getInt(1), rs.getString(2), dateNaissance));
         }
         return abonnes;
     }
 
-    public ArrayList<Document> getDvdsFROMDB(ArrayList<Abonne> abonnes) throws SQLException {
-        ArrayList<Document> dvds = new ArrayList<>();
-        String requete = "Select * from vue_etat_dvd";
-        Statement req2 = this.conn.createStatement();
-        ResultSet rs = req2.executeQuery(requete);
+    public ArrayList<IDocument> getDocumentsFROMDB(ArrayList<Abonne> abonnes) throws SQLException {
+        ArrayList<IDocument> documents = new ArrayList<>();
+        ArrayList<String> documentsType = new ArrayList<>();
+        documentsType.add("dvd");
 
+        for (String type : documentsType) {
+            String request = "Select * from vue_etat_" + type;
+            Statement req2 = this.conn.createStatement();
+            ResultSet rs = req2.executeQuery(request);
+            while(rs.next()) {
+                IDocument document = null;
+                // set etat
+                switch (rs.getString(4)) {
+                    case "LIBRE":
+                        switch (type) {
+                            case "dvd":
+                                document = new DVD(rs.getInt(1), rs.getString(2), Etat.LIBRE, rs.getBoolean(3));
+                                break;
+                        }
+                    break;
 
+                    case "EMPRUNTE":
+                        switch (type) {
+                            case "dvd":
+                                document = new DVD(rs.getInt(1), rs.getString(2), Etat.EMPRUNTE, this.getAbonneById(abonnes,rs.getInt(5)), rs.getBoolean(3));
+                        }
+                    break;
 
+                    case "RESERVE":
+                        switch (type) {
+                            case "dvd":
+                                document = new DVD(rs.getInt(1), rs.getString(2), Etat.RESERVE, this.getAbonneById(abonnes,rs.getInt(6)) , new Date(), rs.getBoolean(3));
+                                break;
+                        }
+                }
 
-        while(rs.next()) {
-            Integer numAb = rs.getInt(5);
-            if(numAb == 0) {    //renvoit 0 quand c'est null en BD
-                numAb = rs.getInt(6);
+                documents.add(document);
             }
-
-            Document dvd = new DVD(rs.getInt(2), rs.getString(1), rs.getBoolean(3), this.getAbonneById(abonnes, numAb), rs.getString(4).toLowerCase());
-            dvds.add(dvd);
         }
-
-        return dvds;
-
+        return documents;
     }
 
     private Abonne getAbonneById(ArrayList<Abonne> abonnes, Integer numero) {
@@ -76,40 +101,34 @@ public class ConnectionBD {
         return null;
     }
 
-    public void reserverDvdBD(Abonne abonne, Document dvd) throws SQLException {
-        Statement req2 = this.conn.createStatement();
-        ResultSet rs = req2.executeQuery("Select idDocument from DVD where idDvd = " + dvd.numero());
-        rs.next();
-        int idDocument = rs.getInt(1);
-
-        //Pour l'instant date null
+    public void reserverDocumentBD(Abonne abonne, IDocument document) throws SQLException {
+        int idDocument = document.numero();
+        Date date = new Date();
         Statement reqInsert = this.conn.createStatement();
-        int rsInsert = req2.executeUpdate("INSERT INTO RESERVATION VALUES (" + abonne.getNumeroAdhérent() + ", " + idDocument + ", null)");
-
-
+        int rsInsert = reqInsert.executeUpdate("INSERT INTO RESERVATION VALUES (" + abonne.getNumeroAdhérent() + ", " + idDocument + ", '" + date + "')");
     }
 
-    public void retournerDvd(Document dvd) throws SQLException {
-        Statement req2 = this.conn.createStatement();
-        ResultSet rs = req2.executeQuery("Select idDocument from DVD where idDvd = " + dvd.numero()); //On récupère l'idDocument du DVD pour l'insérer dans emprunt
-        rs.next();
-        int idDocument = rs.getInt(1);
-
+    public void retournerDocument(IDocument document) throws SQLException {
         Statement req3 = this.conn.createStatement();
-        int rsDelete = req3.executeUpdate("DELETE FROM EMPRUNT WHERE idDocument = " + idDocument);
+        int rsDelete = req3.executeUpdate("DELETE FROM EMPRUNT WHERE idDocument = " + document.numero());
     }
 
-    public void emprunterDvd(Document dvd, Abonne abonne) throws SQLException {
-        Statement req2 = this.conn.createStatement();
-        ResultSet rs = req2.executeQuery("Select idDocument from DVD where idDvd = " + dvd.numero()); //On récupère l'idDocument du DVD pour l'insérer dans emprunt
-        rs.next();
-        int idDocument = rs.getInt(1);
-
-        //on insert le numeroAbonne et l'idDocument dans EMPRUNT
-        Statement req4 = this.conn.createStatement();
-        int rsInsert = req4.executeUpdate("INSERT INTO EMPRUNT VALUES (" + abonne.getNumeroAdhérent() +"," + idDocument + ", null)");
-
-        Statement req5 = this.conn.createStatement();
-        int rsDelete = req5.executeUpdate("DELETE FROM RESERVATION WHERE idDocument = " + idDocument);
+    public void emprunterDocument(IDocument document, Abonne abonne) throws SQLException {
+        int idDocument = document.numero();
+        Date date = new Date();
+        if(document.reservePar() == null || document.reservePar().getNumeroAdhérent() == abonne.getNumeroAdhérent()) {
+            //on insert le numeroAbonne et l'idDocument dans EMPRUNT
+            Statement req4 = this.conn.createStatement();
+            int rsInsert = req4.executeUpdate("INSERT INTO EMPRUNT VALUES (" + abonne.getNumeroAdhérent() +"," + idDocument + ", '" + java.sql.Date.valueOf(java.time.LocalDate.now()) + "')");
+            if (document.reservePar() != null) {
+                //on supprime la reservation
+                Statement req5 = this.conn.createStatement();
+                int rsDelete = req5.executeUpdate("DELETE FROM RESERVATION WHERE idDocument = " + idDocument);
+            }
+            System.out.println("Emprunt effectué");
+        }
+        else {
+            System.out.println("Ce document est réservé par un autre abonné");
+        }
     }
 }
